Add files via upload
The Empire Breakout 2 machine on VulnHub is a Linux-based virtual machine designed for penetration testing and ethical hacking practice. The machine is rated as an easy level CTF challenge, and it requires the user to perform enumeration, exploitation, and privilege escalation techniques to gain access to the system and capture the flags.